To summarize: all "long" and "unsigned long" uses were replaced with proper
"int32_t" and "uint32_t" type uses to be able to produce a 64bit compatible
version of ReGaHss and all hm-apps tools+libraries. Now the "uintXX_t" types
will be used where appropriate so that types are used with specific bit width
rather than general types like "long" which differ from 32bit to 64bit OS
systems. This in fact allows to still be able to compile 32bit versions
without introducing incompatibilities while at the same time allowing
for 64bit versions to be compiled for x86_64 or aarch64.

+++ src/ssdpd/ssdpd.cpp
@@ -31,11 +31,11 @@ static const int REPEAT_AFTER = 1800;
 static std::string serial;
 static std::string uuid;
 
-long long time_millis()
+int64_t time_millis()
 {
   struct timeval te;
   gettimeofday(&te, NULL); // get current time
-  long long milliseconds = te.tv_sec*1000LL + te.tv_usec/1000; // calculate milliseconds
+  int64_t milliseconds = te.tv_sec*1000LL + te.tv_usec/1000; // calculate milliseconds
   return milliseconds;
 }
 
@@ -231,7 +231,7 @@ static void send_udp(int s, struct sockaddr_in* sa, std::string response, int mx
 	if (end != std::string::npos)
 		end += 4;
 	std::set<int>::iterator it = delays.begin();
-	long long t = time_millis();
+	int64_t t = time_millis();
 	while (end != std::string::npos) {
 		std::string msg = response.substr(start, end - start);
 		fetch_uuid(msg);
@@ -239,7 +239,7 @@ static void send_udp(int s, struct sockaddr_in* sa, std::string response, int mx
 		end = response.find("\x0d\x0a\x0d\x0a", start);
 		if (end != std::string::npos)
 			end += 4;
-		long delay = t + (*it) - time_millis();
+		int64_t delay = t + (*it) - time_millis();
 		if (delay > 0)
 			usleep(delay * 1000);
 		int n;
